requirements.txt
pandas
konlpy
toml
discord.py
fastapi~=0.110.0